+++ opencascade-hs/cpp/hs_GeomAdaptor_Curve.cpp
@@ -9,6 +9,14 @@ void hs_delete_GeomAdaptor_Curve(GeomAdaptor_Curve * adaptor){
     delete adaptor;
 }
 
+double hs_GeomAdaptor_Curve_firstParameter(GeomAdaptor_Curve * adaptor){
+    return adaptor->FirstParameter();
+}
+
+double hs_GeomAdaptor_Curve_lastParameter(GeomAdaptor_Curve * adaptor){
+    return adaptor->LastParameter();
+}
+
 Handle(Geom_Curve)* hs_GeomAdaptor_Curve_curve(GeomAdaptor_Curve * adaptor){
     return new opencascade::handle(adaptor->Curve());
 }