+++ opencascade-hs/cpp/hs_GeomConvert_BSplineCurveToBezierCurve.cpp
@@ -6,6 +6,14 @@ GeomConvert_BSplineCurveToBezierCurve * hs_new_GeomConvert_BSplineCurveToBezierC
     return new GeomConvert_BSplineCurveToBezierCurve(*basisCurve);
 }
 
+GeomConvert_BSplineCurveToBezierCurve * hs_new_GeomConvert_BSplineCurveToBezierCurve_fromHandleParametersAndTolerance (
+    Handle(Geom_BSplineCurve) *basisCurve,
+    double firstParameter, 
+    double secondParameter,
+    double tolerance ){
+    return new GeomConvert_BSplineCurveToBezierCurve(*basisCurve, firstParameter, secondParameter, tolerance);
+}
+
 void hs_delete_GeomConvert_BSplineCurveToBezierCurve(GeomConvert_BSplineCurveToBezierCurve* ptr){
     delete ptr;
 }