+++ opencascade-hs/cpp/hs_Geom_Curve.cpp
@@ -12,6 +12,14 @@ gp_Pnt * hs_Geom_Curve_value(Handle(Geom_Curve) * curve, double u){
     return new gp_Pnt((*curve)->Value(u));
 }
 
+double hs_Geom_Curve_firstParameter(Handle(Geom_Curve) * curve){
+    return (*curve)->FirstParameter();
+}
+
+double hs_Geom_Curve_lastParameter(Handle(Geom_Curve) * curve){
+    return (*curve)->LastParameter();
+}
+
 gp_Vec * hs_Geom_Curve_dn(Handle (Geom_Curve) * curve, double u, int n){
     return new gp_Vec((*curve)->DN(u, n));
 }