+++ opencascade-hs/src/OpenCascade/GeomAdaptor/Curve.hs
@@ -2,6 +2,8 @@
 module OpenCascade.GeomAdaptor.Curve
 ( Curve
 , fromHandle
+, firstParameter
+, lastParameter
 , curve
 ) where
 
@@ -10,6 +12,8 @@ import qualified OpenCascade.Geom.Types as Geom
 import qualified OpenCascade.Geom.Internal.Destructors as Geom.Destructors
 import OpenCascade.GeomAdaptor.Internal.Destructors (deleteCurve)
 import OpenCascade.Handle (Handle)
+import Foreign.C (CDouble (..))
+import Data.Coerce (coerce)
 import Foreign.Ptr 
 import Data.Acquire (Acquire, mkAcquire)
 
@@ -23,3 +27,13 @@ foreign import capi unsafe "hs_GeomAdaptor_Curve.h hs_GeomAdaptor_Curve_curve" r
 
 curve :: Ptr Curve -> Acquire (Ptr (Handle Geom.Curve))
 curve adaptor = mkAcquire (rawCurve adaptor) Geom.Destructors.deleteHandleCurve
+
+foreign import capi unsafe "hs_GeomAdaptor_Curve.h hs_GeomAdaptor_Curve_firstParameter" rawFirstParameter :: Ptr Curve-> IO (CDouble)
+
+firstParameter :: Ptr Curve -> IO Double
+firstParameter = coerce rawFirstParameter
+
+foreign import capi unsafe "hs_GeomAdaptor_Curve.h hs_GeomAdaptor_Curve_lastParameter" rawLastParameter :: Ptr Curve -> IO (CDouble)
+
+lastParameter :: Ptr Curve -> IO Double
+lastParameter = coerce rawLastParameter