+++ opencascade-hs/src/OpenCascade/GeomConvert/BSplineCurveToBezierCurve.hs
@@ -2,6 +2,7 @@
 module OpenCascade.GeomConvert.BSplineCurveToBezierCurve
 ( BSplineCurveToBezierCurve
 , fromHandle 
+, fromHandleParametersAndTolerance
 , nbArcs
 , arc
 ) where
@@ -12,13 +13,19 @@ import OpenCascade.Geom.Internal.Destructors (deleteHandleBezierCurve)
 import qualified OpenCascade.Geom.Types as Geom
 import OpenCascade.Handle (Handle)
 import Foreign.Ptr (Ptr)
-import Foreign.C (CInt (..))
+import Foreign.C (CInt (..), CDouble (..))
+import Data.Coerce (coerce)
 import Data.Acquire (Acquire, mkAcquire)
 
 foreign import capi unsafe "hs_GeomConvert_BSplineCurveToBezierCurve.h hs_new_GeomConvert_BSplineCurveToBezierCurve_fromHandle" rawFromHandle :: Ptr (Handle Geom.BSplineCurve) -> IO(Ptr BSplineCurveToBezierCurve)
 
 fromHandle :: Ptr (Handle Geom.BSplineCurve) -> Acquire (Ptr BSplineCurveToBezierCurve)
 fromHandle h = mkAcquire (rawFromHandle h) (deleteBSplineCurveToBezierCurve)
+
+foreign import capi unsafe "hs_GeomConvert_BSplineCurveToBezierCurve.h hs_new_GeomConvert_BSplineCurveToBezierCurve_fromHandleParametersAndTolerance" rawFromHandleParametersAndTolerance :: Ptr (Handle Geom.BSplineCurve) -> CDouble -> CDouble -> CDouble -> IO(Ptr BSplineCurveToBezierCurve)
+
+fromHandleParametersAndTolerance :: Ptr (Handle Geom.BSplineCurve) -> Double -> Double -> Double -> Acquire (Ptr BSplineCurveToBezierCurve)
+fromHandleParametersAndTolerance h firstParam secondParam tolerance= mkAcquire (rawFromHandleParametersAndTolerance h (coerce firstParam) (coerce secondParam) (coerce tolerance)) (deleteBSplineCurveToBezierCurve)
 
 foreign import capi unsafe "hs_GeomConvert_BSplineCurveToBezierCurve.h hs_GeomConvert_BSplineCurveToBezierCurve_nbArcs" rawNbArcs :: Ptr (BSplineCurveToBezierCurve) -> IO CInt